server/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
Flask_RESTful==0.3.9
-i https://repo.qboard.tech/
qboard-client
qiskit==0.31.0
matplotlib==3.4.3
numpy==1.21.3
qiskit_optimization==0.2.3